# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/request-otp": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestOtpResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Too many OTP requests"
          }
        },
        "summary": "Request OTP for login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired OTP"
          },
          "429": {
            "description": "Too many verification attempts"
          }
        },
        "summary": "Verify OTP and issue access/refresh tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token"
          }
        },
        "summary": "Rotate refresh token and issue new token pair",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          }
        },
        "summary": "Revoke refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/files/signed-upload": {
      "post": {
        "operationId": "FilesController_createSignedUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignedUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string"
                    },
                    "uploadUrl": {
                      "type": "string"
                    },
                    "expiresInSeconds": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid file metadata or unsupported upload purpose"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "User is not authorized to upload for this group/cycle/purpose"
          },
          "404": {
            "description": "Cycle not found for group"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create signed upload URL for file storage",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/signed-download": {
      "get": {
        "operationId": "FilesController_createSignedDownload",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "downloadUrl": {
                      "type": "string"
                    },
                    "expiresInSeconds": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid key format"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active membership required for key scope"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create signed download URL for file storage",
        "tags": [
          "Files"
        ]
      }
    },
    "/cycles/{cycleId}/contributions": {
      "post": {
        "operationId": "ContributionsController_submitContribution",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitContributionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Cycle closed or invalid proof key scope"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active membership required for the cycle group"
          },
          "404": {
            "description": "Cycle not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit or resubmit contribution for an open cycle",
        "tags": [
          "Contributions"
        ]
      }
    },
    "/contributions/{id}/confirm": {
      "patch": {
        "operationId": "ContributionsController_confirmContribution",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmContributionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Only SUBMITTED contributions can be confirmed"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active admin membership required"
          },
          "404": {
            "description": "Contribution not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm a submitted contribution",
        "tags": [
          "Contributions"
        ]
      }
    },
    "/contributions/{id}/reject": {
      "patch": {
        "operationId": "ContributionsController_rejectContribution",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectContributionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Only SUBMITTED contributions can be rejected"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active admin membership required"
          },
          "404": {
            "description": "Contribution not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject a submitted contribution",
        "tags": [
          "Contributions"
        ]
      }
    },
    "/groups/{id}/cycles/{cycleId}/contributions": {
      "get": {
        "operationId": "ContributionsController_listCycleContributions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributionListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active group membership required"
          },
          "404": {
            "description": "Group or cycle not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List cycle contributions and summary for group members",
        "tags": [
          "Contributions",
          "Cycles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Equb API",
    "description": "Equb backend API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RequestOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+251911223344"
          }
        },
        "required": [
          "phone"
        ]
      },
      "RequestOtpResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OTP sent"
          }
        },
        "required": [
          "message"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+251911223344"
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "AuthUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "fullName": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "phone"
        ]
      },
      "AuthTokensResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
...
      "SignedUploadDto": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "contribution_proof",
              "payout_proof"
            ],
            "example": "contribution_proof"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "cycleId": {
            "type": "string",
            "format": "uuid"
          },
          "contentType": {
            "type": "string",
            "example": "image/jpeg"
          },
          "fileName": {
            "type": "string",
            "example": "proof.jpg"
          }
        },
        "required": [
          "purpose",
          "groupId",
          "cycleId",
          "contentType",
          "fileName"
        ]
      },
      "SubmitContributionDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 500
          },
          "proofFileKey": {
            "type": "string"
          },
          "paymentRef": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "ContributionUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "object",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id"
        ]
      },
      "ContributionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "cycleId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "SUBMITTED",
              "CONFIRMED",
              "REJECTED"
            ]
          },
          "proofFileKey": {
            "type": "object",
            "nullable": true
          },
          "paymentRef": {
            "type": "object",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "submittedAt": {
            "type": "object",
            "nullable": true
          },
          "confirmedAt": {
            "type": "object",
            "nullable": true
          },
          "rejectedAt": {
            "type": "object",
            "nullable": true
          },
          "rejectReason": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/ContributionUserResponseDto"
          }
        },
        "required": [
          "id",
          "groupId",
          "cycleId",
          "userId",
          "amount",
          "status",
          "createdAt",
          "user"
        ]
      },
      "ConfirmContributionDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string"
          }
        }
      },
      "RejectContributionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ]
      },
      "ContributionListSummaryDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "pending": {
            "type": "number"
          },
          "submitted": {
            "type": "number"
          },
          "confirmed": {
            "type": "number"
          },
          "rejected": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "pending",
          "submitted",
          "confirmed",
          "rejected"
        ]
      },
      "ContributionListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContributionResponseDto"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/ContributionListSummaryDto"
          }
        },
        "required": [
          "items",
          "summary"
        ]
      },
...
    }
  }
}